backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
geopandas>=0.14.4
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
joblib>=1.4.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
shapely>=2.0.4
openpyxl>=3.1.2
xlsxwriter>=3.2.0
python-multipart>=0.0.9
